Fix links to non-html files
Because of this issue:

  https://github.com/GitbookIO/gitbook/issues/1539

links to internal resources that are not HTML/MD seem to be
failing.  Clicking on them in a browser causes no discernible
behavior.  In the JavaScript console, it becomes clear
that the resource is being read and attempted to be parsed
as HTML, instead of served up as a file.

As a workaround, this commit uses explicit anchor tags
with targets and download attributes.