Update modules.ml
It's very possible I'm misunderstanding something, but I think the word "subset" may not be correct in this section. If S1, S2, etc are subtypes of T, then the module values they contain should be supersets of T (as per the rule above). I don't think the set of module values of S1, S2 etc should be contained in the mod. values of T. Let me know if I'm misunderstanding!