Make OCaml Platform installation clearer
From the 3110 slack: "Could we fix the textbook to make the OCaml Platform extension more obvious? As is, it's a bullet underneath the Mac instructions and lots of people have not seen to install it (the recent ed post and at least three people during my install clinic hours yesterday)"

I've proposed the following changes to the book which should make this clearer, and added an image to draw attention to this bullet point, as well as make the warning about the outdated extensions obvious.