Build without Topfind
The textbook previously implicitly assumed that `Topfind` was
being loaded by `ocaml`. Though that is often the case (for
example, if the user has installed `user-setup` in OPAM), it
need not always be true. In fact it is arguably better not
to have to always load `Topfind`, because in-class demos
with `utop` experience excessive output when it is loaded.

This commit adds a `#use "topfind"` before each `#require` so that
`Topfind` is quietly loaded by the textbook. The output is hidden
from the HTML.