Update discussion of polymorphic functions
OCaml used to understand the type in `let x : 'a -> 'a` to mean
`'a . 'a -> 'a`. However in the last year it seems the universal
quantification has been dropped, and OCaml permits the type variable
to be instantiated at the time of let binding.

This commit addresses that change, though it's unsatisfying that
we now have to remove a negative example. The only other fix
I can imagine would be to introduce the universal type quantification
syntax, but that's a step farther than I'm willing to make for now.