Update book logo
This is a trial of a new logo. Previously we used the "lambda
of power", which was an in-joke from the AY2020-21 offering
of CS 3110. The original graphic asset for it has since been
lost.

The new logo is based on the initial of the book, OP. It features
a lotus flower highlighted in orange. The color is a nod to OCaml.
The flower is a nod to beauty (a subtitle of the book). The lotus
flower in particular is a nod to enlightenment (surely an effect
of functional programming ;)) and purity (because we avoid mutability
as long as possible in the book).

The logo could be updated in the future based on community feedback.