Expose the exception 'Empty' as part of the module signature.
The given module provides both versions of the interface functions: one
that raises an exception and one that returns an option. This exception
should therefore be exposed in the module signature. (This is how it was
done in the previous examples, and also how it is done in the standard
library, e.g., https://v2.ocaml.org/api/type_Stack.html.)