Fix a bug in [Promise.bind]
Our implementation of [Promise.bind] violated the specification
of that function: given [bind p f], if [p] was already fulfilled
and [f] raised an exception, we previously had [bind] raise that
exception instead of having [bind] return a rejected promise
containing that exception.

This commit fixes it by having [bind] catch the exception and
construct the rejected promise to return.

Intriguingly Lwt had the same bug but chose not to fix it.